+++ include/world_builder/features/continental_plate_models/composition/interface.h
@@ -22,8 +22,7 @@
 
 
 #include "world_builder/parameters.h"
-#include "world_builder/utilities.h"
-
+#include "world_builder/objects/natural_coordinate.h"
 
 namespace WorldBuilder
 {
@@ -75,7 +74,7 @@ namespace WorldBuilder
              */
             virtual
             double get_composition(const Point<3> &position,
-                                   const WorldBuilder::Utilities::NaturalCoordinate &position_in_natural_coordinates,
+                                   const Objects::NaturalCoordinate &position_in_natural_coordinates,
                                    const double depth,
                                    const unsigned int composition_number,
                                    double composition,