+++ include/world_builder/features/continental_plate_models/grains/uniform.h
@@ -86,7 +86,7 @@ namespace WorldBuilder
              */
             WorldBuilder::grains
             get_grains(const Point<3> &position,
-                       const WorldBuilder::Utilities::NaturalCoordinate &position_in_natural_coordinates,
+                       const Objects::NaturalCoordinate &position_in_natural_coordinates,
                        const double depth,
                        const unsigned int composition_number,
                        WorldBuilder::grains grains,