+++ include/world_builder/features/continental_plate_models/temperature/uniform.h
@@ -24,7 +24,6 @@
 #include "world_builder/features/continental_plate_models/temperature/interface.h"
 #include "world_builder/features/utilities.h"
 #include "world_builder/objects/surface.h"
-#include "world_builder/utilities.h"
 
 
 namespace WorldBuilder
@@ -72,7 +71,7 @@ namespace WorldBuilder
              * gravity and current temperature.
              */
             double get_temperature(const Point<3> &position,
-                                   const WorldBuilder::Utilities::NaturalCoordinate &position_in_natural_coordinates,
+                                   const Objects::NaturalCoordinate &position_in_natural_coordinates,
                                    const double depth,
                                    const double gravity,
                                    double temperature,