+++ include/world_builder/utilities.h
@@ -23,6 +23,7 @@
 
 #include "world_builder/nan.h"
 #include "world_builder/coordinate_systems/interface.h"
+#include "world_builder/objects/natural_coordinate.h"
 
 
 namespace WorldBuilder
@@ -79,72 +80,6 @@ namespace WorldBuilder
     signed_distance_to_polygon(const std::vector<Point<2> > &point_list_,
                                const Point<2> &point_);
 
-
-    /*
-    * A class that represents a point in a chosen coordinate system.
-    */
-    class NaturalCoordinate
-    {
-      public:
-        /**
-         * Constructor based on providing the geometry model as a pointer
-         */
-        NaturalCoordinate(const std::array<double,3> &position,
-                          const ::WorldBuilder::CoordinateSystems::Interface &coordinate_system);
-
-        /**
-         * Constructor based on providing the geometry model as a pointer
-         */
-        NaturalCoordinate(const Point<3> &position,
-                          const ::WorldBuilder::CoordinateSystems::Interface &coordinate_system);
-
-        /**
-         * Returns the coordinates in the given coordinate system, which may
-         * not be Cartesian.
-         */
-        const std::array<double,3> &get_coordinates() const;
-
-        /**
-         * The coordinate that represents the 'surface' directions in the
-         * chosen coordinate system.
-         */
-        std::array<double,2> get_surface_coordinates() const;
-
-        /**
-         * The coordinate that represents the 'surface' directions in the
-         * chosen coordinate system.
-         */
-        Point<2> get_surface_point() const;
-
-        /**
-         * The coordinate that represents the 'depth' direction in the chosen
-         * coordinate system.
-         */
-        double get_depth_coordinate() const;
-
-        /**
-         * Return a reference to the coordinate that represents the 'depth' direction
-         * in the chosen coordinate system.
-         */
-        double &get_ref_depth_coordinate();
-
-        /**
-         * get the coordinate system type of this coordinate.
-         */
-        CoordinateSystem get_coordinate_system() const;
-
-      private:
-        /**
-         * An enum which stores the the coordinate system of this natural
-         * point
-         */
-        CoordinateSystem coordinate_system;
-
-        /**
-         * An array which stores the coordinates in the coordinates system
-         */
-        std::array<double,3> coordinates;
-    };
 
     /**
      * Returns spherical coordinates of a Cartesian point. The returned array
@@ -404,7 +339,7 @@ namespace WorldBuilder
      * and the average angle of the closest segment/section.
      */
     PointDistanceFromCurvedPlanes distance_point_from_curved_planes(const Point<3> &check_point,
-                                                                    const NaturalCoordinate &check_point_natural,
+                                                                    const Objects::NaturalCoordinate &check_point_natural,
                                                                     const Point<2> &reference_point,
                                                                     const std::vector<Point<2> > &point_list,
                                                                     const std::vector<std::vector<double> > &plane_segment_lengths,