+++ source/features/continental_plate.cc
@@ -145,7 +145,7 @@ namespace WorldBuilder
 
     double
     ContinentalPlate::temperature(const Point<3> &position_in_cartesian_coordinates,
-                                  const NaturalCoordinate &position_in_natural_coordinates,
+                                  const Objects::NaturalCoordinate &position_in_natural_coordinates,
                                   const double depth,
                                   const double gravity_norm,
                                   double temperature) const
@@ -182,7 +182,7 @@ namespace WorldBuilder
 
     double
     ContinentalPlate::composition(const Point<3> &position_in_cartesian_coordinates,
-                                  const NaturalCoordinate &position_in_natural_coordinates,
+                                  const Objects::NaturalCoordinate &position_in_natural_coordinates,
                                   const double depth,
                                   const unsigned int composition_number,
                                   double composition) const
@@ -220,7 +220,7 @@ namespace WorldBuilder
 
     WorldBuilder::grains
     ContinentalPlate::grains(const Point<3> &position_in_cartesian_coordinates,
-                             const NaturalCoordinate &position_in_natural_coordinates,
+                             const Objects::NaturalCoordinate &position_in_natural_coordinates,
                              const double depth,
                              const unsigned int composition_number,
                              WorldBuilder::grains grains) const