+++ source/features/continental_plate_models/composition/uniform.cc
@@ -103,7 +103,7 @@ namespace WorldBuilder
 
         double
         Uniform::get_composition(const Point<3> & /*position_in_cartesian_coordinates*/,
-                                 const NaturalCoordinate &position_in_natural_coordinates,
+                                 const Objects::NaturalCoordinate &position_in_natural_coordinates,
                                  const double depth,
                                  const unsigned int composition_number,
                                  double composition_,