+++ source/features/continental_plate_models/temperature/linear.cc
@@ -97,7 +97,7 @@ namespace WorldBuilder
 
         double
         Linear::get_temperature(const Point<3> & /*position_in_cartesian_coordinates*/,
-                                const WorldBuilder::Utilities::NaturalCoordinate &position_in_natural_coordinates,
+                                const Objects::NaturalCoordinate &position_in_natural_coordinates,
                                 const double depth,
                                 const double gravity_norm,
                                 double temperature_,