+++ source/features/oceanic_plate_models/temperature/plate_model_constant_age.cc
@@ -100,7 +100,7 @@ namespace WorldBuilder
 
         double
         PlateModelConstantAge::get_temperature(const Point<3> & /*position*/,
-                                               const WorldBuilder::Utilities::NaturalCoordinate &position_in_natural_coordinates,
+                                               const Objects::NaturalCoordinate &position_in_natural_coordinates,
                                                const double depth,
                                                const double gravity_norm,
                                                double temperature_,