+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Changed the name of the main branch from master to main \[Menno Fraters; 2021-10-28; [#350](https://github.com/GeodynamicWorldBuilder/WorldBuilder/pull/350)\]
 - Ridge coordinates are now an array of ridges, allowing mulitple ridges within a single oceanic plate with transform faults in between. \[Menno Fraters; 2021-11-03; [#362](https://github.com/GeodynamicWorldBuilder/WorldBuilder/pull/362)\]
 - parse entries for all temperature, composition and grain plugins now require the coordinates of the feature to be passed as a parameter. \[Menno Fraters; 2022-03-26; [#396](https://github.com/GeodynamicWorldBuilder/WorldBuilder/pull/396)\]
+- NaturalCoordinate has moved out of utilities to the objects namespace and folder. \[Menno Fraters; 2022-03-26; [#399](https://github.com/GeodynamicWorldBuilder/WorldBuilder/pull/399)\]
 
 ### Fixed
 - Using slabs and faults beyond the -180 to 180 range gave issues. These are now fixed and it now works and is tested for the -380 to 380 range. \[Menno Fraters; 2021-10-22; [#338](https://github.com/GeodynamicWorldBuilder/WorldBuilder/issues/338), [#340](https://github.com/GeodynamicWorldBuilder/WorldBuilder/pull/340) and [#342](https://github.com/GeodynamicWorldBuilder/WorldBuilder/pull/342)\]